% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-utils.R
\name{dp}
\alias{dp}
\title{Specify decimal places for printing}
\usage{
dp(n)
}
\arguments{
\item{n}{Number of decimal places to print. Use negative values for positive
powers of 10; e.g. \code{n = -2} will print values rounded to the nearest
multiple of 100.}
}
\value{
A vector of the same length as \code{n}, suitable to be passed to the
\code{accuracy} argument of \code{\link{nmbr}()} and related functions.
}
\description{
This function specifies the number of decimal places to print (when\
passed to the \code{accuracy} argument of \code{\link{nmbr}()} and
related functions).
}
\seealso{
\code{\link{sf}}, \code{\link{number-formatting}}
}
