% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-utils.R
\name{sf}
\alias{sf}
\title{Specify significant figures for printing}
\usage{
sf(x, n, dp_min = -Inf, dp_max = Inf)
}
\arguments{
\item{x}{Vector of values for which significant figures are calculated}

\item{n}{Number of significant figures to print}

\item{dp_min}{(optional) Minimum number of decimal places to print. Use
negative values for positive powers of 10; e.g., \code{dp_min = -2} will
print all values to at least the 100's digit, even if that results in
more than \code{n} significant figures.}

\item{dp_max}{(optional) Maximum number of decimal places to print. Use
negative values for positive powers of 10; e.g., \code{dp_max = -2} will
print all values to at most the 100's digit, even if that results in
fewer than \code{n} significant figures.}
}
\value{
A vector of the same length as \code{x}, suitable to be passed to the
\code{accuracy} argument of \code{\link{nmbr}()} and related functions.
}
\description{
This function specifies the number of significant figures to print (when\
passed to the \code{accuracy} argument of \code{\link{nmbr}()} and
related functions).
}
\seealso{
\code{\link{dp}}, \code{\link{number-formatting}}
}
