% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-values.R
\name{sstars}
\alias{sstars}
\title{Print 'significance' stars for given P-value statistics}
\usage{
sstars(p, levels = c(`*` = 0.1, `**` = 0.05, `***` = 0.01))
}
\arguments{
\item{p}{Vector of p-values.}

\item{levels}{Named vector of significance levels. The values should be the
upper threshold of each level, and the names the associated labels to be
printed.}
}
\description{
Print significance stars representing p < 0.01, p < 0.05, p < 0.1 or custom
significance levels. The usual caveats about the misuse of p-values
apply - these should not be taken to represent the correctness or
otherwise of a given alternative hypothesis.
}
