% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmbr.R
\name{number-formatting}
\alias{number-formatting}
\alias{nmbr}
\alias{prct}
\alias{cmma}
\alias{dllr}
\alias{pval}
\alias{create_nmbr}
\title{Format numbers using \code{scales::number()}-type functions}
\usage{
nmbr(
  x,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)

prct(
  x,
  percent = "\%",
  accuracy = 1,
  prefix = "",
  big.mark = "< >",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)

cmma(
  x,
  comma = ",",
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)

dllr(
  x,
  dollar = "$",
  accuracy = 1,
  scale = 1,
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)

pval(
  x,
  accuracy = 1e-04,
  min_p = accuracy,
  add_p = FALSE,
  decimal.mark = ".",
  na = NA_character_,
  ...
)

create_nmbr(
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)
}
\arguments{
\item{x}{Numeric vector to format}

\item{accuracy, scale, prefix, suffix, big.mark, decimal.mark, ...}{As in
\code{\link[scales]{number}}. If a vector is supplied, will be applied
element-wise to \code{x} (and must have the same length as \code{x}).}

\item{html}{Logical scalar. Whether to include formatting marks (minus
signs and narrow spaces between digits) as HTML strings (the default;
best for Word or HTML output documents) or unicode.}

\item{na}{String scalar, replacement to use for missing values in \code{x}.}

\item{percent, comma, dollar}{String scalar to use for the specific formatting
method (percent sign, comma separator, dollar sign, etc).}

\item{min_p}{Numeric scalar. The smallest p-value to print; values smaller
than this will be printed as \code{"<`min_p`".}}

\item{add_p}{Logical scalar. Should \verb{p=} be included before formatted
p-values?}
}
\value{
For \code{create_nmbr()}, a function with the same arguments as \code{nmbr()};
for other functions a character string applying the specified formatting
rules to \code{x}.
}
\description{
These functions extend the \code{\link[scales]{number}}-type formatting
functions, with nice printing of negative numbers, optional replacement
of missing values, and vectorised formatting options.
}
