% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmbr.R
\name{number-formatting}
\alias{number-formatting}
\alias{nmbr}
\alias{prct}
\alias{cmma}
\alias{dllr}
\title{Format numbers using \code{scales::number()}-type functions}
\usage{
nmbr(
  x,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)

prct(
  x,
  symbol = "\%",
  accuracy = 1,
  prefix = "",
  big.mark = "< >",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)

cmma(
  x,
  symbol = ",",
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)

dllr(
  x,
  symbol = "$",
  accuracy = 1,
  scale = 1,
  suffix = "",
  big.mark = "< >",
  decimal.mark = ".",
  html = TRUE,
  na = NA_character_,
  ...
)
}
\arguments{
\item{x}{Numeric vector to format}

\item{accuracy, scale, prefix, suffix, big.mark, decimal.mark, ...}{As in
\code{\link[scales]{number}}. If a vector is supplied, will be applied
element-wise to \code{x} (and must have the same length as \code{x}).}

\item{html}{Logical scalar. Whether to include formatting marks (minus
signs and narrow spaces between digits) as HTML strings (the default;
best for Word or HTML output documents) or unicode.}

\item{na}{String scalar, replacement to use for missing values in \code{x}.}

\item{symbol}{String scalar, value to use for the unit measure of specific
formatting methods (percent sign, dollar sign, etc)}
}
\description{
These functions extend the \code{scales::number()-type} formatting functions,
with nice printing of negative numbers, optional replacement of missing
values, and vectorised formatting options.
}
