% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frmt.R
\name{output-formatting}
\alias{output-formatting}
\alias{mean_sd}
\alias{n_percent}
\alias{est_ci}
\title{Results formatting for tables and reports}
\usage{
mean_sd(mean, sd, prefix = "", suffix = "", ..., format_paren = TRUE)

n_percent(
  n,
  proportion,
  accuracy = 1,
  scale = 1,
  prefix = "",
  suffix = "",
  percent = "\%",
  ...,
  format_paren = TRUE
)

est_ci(
  est,
  low,
  high,
  prefix = "",
  suffix = "",
  ...,
  .sep = " to ",
  format_paren = TRUE
)
}
\arguments{
\item{mean, sd, n, proportion, est, low, high}{Numeric vectors. If \code{proportion} is
missing, it will be calculated as \code{n / sum(n)}.}

\item{format_paren}{Logical scalar. Should the values inside parentheses
include the specified prefix and suffix formatters?}

\item{accuracy, scale, prefix, suffix, percent, ...}{Passed to \code{nmbr()} and
\code{prct()}.}

\item{.sep}{Separator between lower and upper bounds of the CI; default is
\code{' to '}.}
}
\description{
These functions are simple wrappers around \code{paste()} and \code{nmbr()} to format
results for printing.
}
